refactor: rename pakages of ui and organize them
