+++ design_system/src/main/res/drawable/filter_horizontal.xml
@@ -1,0 +1,24 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="20dp"
+    android:height="20dp"
+    android:viewportWidth="20"
+    android:viewportHeight="20">
+  <path
+      android:pathData="M7.479,2.709H7.521C7.891,2.709 8.201,2.709 8.454,2.726C8.719,2.744 8.968,2.783 9.21,2.883C9.772,3.116 10.218,3.562 10.451,4.124C10.551,4.366 10.59,4.615 10.608,4.88C10.625,5.133 10.625,5.443 10.625,5.813V5.855C10.625,6.225 10.625,6.535 10.608,6.788C10.59,7.053 10.551,7.302 10.451,7.544C10.218,8.106 9.772,8.552 9.21,8.785C8.968,8.885 8.719,8.924 8.454,8.942C8.201,8.959 7.891,8.959 7.521,8.959H7.479C7.109,8.959 6.799,8.959 6.546,8.942C6.281,8.924 6.032,8.885 5.79,8.785C5.228,8.552 4.782,8.106 4.549,7.544C4.449,7.302 4.41,7.053 4.392,6.788C4.385,6.688 4.381,6.578 4.379,6.459H2.5C2.155,6.459 1.875,6.179 1.875,5.834C1.875,5.489 2.155,5.209 2.5,5.209H4.379C4.381,5.09 4.385,4.98 4.392,4.88C4.41,4.615 4.449,4.366 4.549,4.124C4.782,3.562 5.228,3.116 5.79,2.883C6.032,2.783 6.281,2.744 6.546,2.726C6.799,2.709 7.109,2.709 7.479,2.709ZM6.631,3.973C6.43,3.987 6.332,4.012 6.268,4.038C6.013,4.144 5.81,4.347 5.704,4.602C5.678,4.666 5.653,4.764 5.639,4.965C5.625,5.171 5.625,5.437 5.625,5.834C5.625,6.231 5.625,6.497 5.639,6.703C5.653,6.904 5.678,7.002 5.704,7.066C5.81,7.321 6.013,7.524 6.268,7.63C6.332,7.656 6.43,7.681 6.631,7.695C6.837,7.709 7.103,7.709 7.5,7.709C7.897,7.709 8.163,7.709 8.369,7.695C8.57,7.681 8.668,7.656 8.732,7.63C8.987,7.524 9.19,7.321 9.296,7.066C9.322,7.002 9.347,6.904 9.361,6.703C9.375,6.497 9.375,6.231 9.375,5.834C9.375,5.437 9.375,5.171 9.361,4.965C9.347,4.764 9.322,4.666 9.296,4.602C9.19,4.347 8.987,4.144 8.732,4.038C8.668,4.012 8.57,3.987 8.369,3.973C8.163,3.959 7.897,3.959 7.5,3.959C7.103,3.959 6.837,3.959 6.631,3.973Z"
+      android:fillColor="#ffffff"
+      android:fillAlpha="0.87"
+      android:fillType="evenOdd"/>
+  <path
+      android:pathData="M2.5,13.542C2.155,13.542 1.875,13.822 1.875,14.167C1.875,14.512 2.155,14.792 2.5,14.792H7.5C7.845,14.792 8.125,14.512 8.125,14.167C8.125,13.822 7.845,13.542 7.5,13.542H2.5Z"
+      android:fillColor="#ffffff"
+      android:fillAlpha="0.87"/>
+  <path
+      android:pathData="M15.621,13.542C15.619,13.424 15.615,13.314 15.608,13.213C15.59,12.949 15.551,12.699 15.451,12.457C15.218,11.896 14.772,11.449 14.21,11.217C13.968,11.116 13.719,11.078 13.455,11.06C13.201,11.042 12.892,11.042 12.521,11.042H12.479C12.109,11.042 11.799,11.042 11.545,11.06C11.281,11.078 11.032,11.116 10.79,11.217C10.228,11.449 9.782,11.896 9.549,12.457C9.449,12.699 9.41,12.949 9.392,13.213C9.375,13.467 9.375,13.776 9.375,14.146V14.188C9.375,14.559 9.375,14.868 9.392,15.122C9.41,15.386 9.449,15.636 9.549,15.878C9.782,16.439 10.228,16.885 10.79,17.118C11.032,17.218 11.281,17.257 11.545,17.275C11.799,17.292 12.108,17.292 12.479,17.292H12.521C12.891,17.292 13.201,17.292 13.455,17.275C13.719,17.257 13.968,17.218 14.21,17.118C14.772,16.885 15.218,16.439 15.451,15.878C15.551,15.636 15.59,15.386 15.608,15.122C15.615,15.021 15.619,14.911 15.621,14.792H17.5C17.845,14.792 18.125,14.512 18.125,14.167C18.125,13.822 17.845,13.542 17.5,13.542H15.621ZM14.375,14.167C14.375,13.771 14.375,13.504 14.361,13.298C14.347,13.097 14.322,13 14.296,12.935C14.19,12.68 13.987,12.477 13.732,12.372C13.668,12.345 13.57,12.32 13.369,12.307C13.163,12.293 12.897,12.292 12.5,12.292C12.103,12.292 11.837,12.293 11.631,12.307C11.43,12.32 11.332,12.345 11.268,12.372C11.013,12.477 10.81,12.68 10.704,12.935C10.678,13 10.653,13.097 10.639,13.298C10.625,13.504 10.625,13.771 10.625,14.167C10.625,14.564 10.625,14.83 10.639,15.037C10.653,15.237 10.678,15.335 10.704,15.399C10.81,15.655 11.013,15.857 11.268,15.963C11.332,15.99 11.43,16.014 11.631,16.028C11.837,16.042 12.103,16.042 12.5,16.042C12.897,16.042 13.163,16.042 13.369,16.028C13.57,16.014 13.668,15.99 13.732,15.963C13.987,15.857 14.19,15.655 14.296,15.399C14.322,15.335 14.347,15.237 14.361,15.037C14.375,14.83 14.375,14.564 14.375,14.167Z"
+      android:fillColor="#ffffff"
+      android:fillAlpha="0.87"
+      android:fillType="evenOdd"/>
+  <path
+      android:pathData="M12.5,5.209C12.155,5.209 11.875,5.489 11.875,5.834C11.875,6.179 12.155,6.459 12.5,6.459L17.5,6.459C17.845,6.459 18.125,6.179 18.125,5.834C18.125,5.489 17.845,5.209 17.5,5.209L12.5,5.209Z"
+      android:fillColor="#ffffff"
+      android:fillAlpha="0.87"/>
+</vector>
