+++ design_system/src/main/res/drawable/search_icon.xml
@@ -1,0 +1,12 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M11.5,21.75C5.85,21.75 1.25,17.15 1.25,11.5C1.25,5.85 5.85,1.25 11.5,1.25C17.15,1.25 21.75,5.85 21.75,11.5C21.75,17.15 17.15,21.75 11.5,21.75ZM11.5,2.75C6.67,2.75 2.75,6.68 2.75,11.5C2.75,16.32 6.67,20.25 11.5,20.25C16.33,20.25 20.25,16.32 20.25,11.5C20.25,6.68 16.33,2.75 11.5,2.75Z"
+      android:fillColor="#969799"/>
+  <path
+      android:pathData="M22,22.75C21.81,22.75 21.62,22.68 21.47,22.529L19.47,20.529C19.18,20.24 19.18,19.76 19.47,19.469C19.76,19.18 20.24,19.18 20.53,19.469L22.53,21.469C22.82,21.76 22.82,22.24 22.53,22.529C22.38,22.68 22.19,22.75 22,22.75Z"
+      android:fillColor="#969799"/>
+</vector>
