+++ ui/src/main/java/com/baghdad/feature/search/SearchScreen.kt
@@ -1,4 +1,4 @@
-package com.baghdad.screen
+package com.baghdad.feature.search
 
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
@@ -7,8 +7,8 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.baghdad.component.RecentSearch
-import com.baghdad.component.RecentViewedSection
+import com.baghdad.feature.search.component.RecentSearch
+import com.baghdad.feature.search.component.RecentViewedSection
 import com.baghdad.viewmodel.search.SearchInteractionListener
 import com.baghdad.viewmodel.search.SearchScreenState
