+++ ui/src/main/java/com/baghdad/feature/search/component/RecentViewed.kt
@@ -1,4 +1,4 @@
-package com.baghdad.component
+package com.baghdad.feature.search.component
 
 import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -7,6 +7,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
+import com.baghdad.component.MoviesCard
 import com.baghdad.ui.R
 import com.baghdad.viewmodel.search.SearchInteractionListener
 import com.baghdad.viewmodel.search.SearchScreenState
